requirements.txt
aiofiles==0.4.0
aniso8601==7.0.0
async-exit-stack==1.0.1
async-generator==1.10
bcrypt==3.1.7
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
dnspython==1.16.0
email-validator==1.0.5
fastapi==0.52.0
graphene==2.1.8
graphql-core==2.3.1
graphql-relay==2.0.1
h11==0.9.0
httptools==0.1.1
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
passlib==1.7.2
promise==2.3
psycopg2-binary==2.8.4
pycparser==2.20
pydantic==1.4
PyJWT==1.7.1
python-multipart==0.0.5
PyYAML==5.3
requests==2.23.0
Rx==1.6.1
six==1.14.0
SQLAlchemy==1.3.13
starlette==0.13.2
ujson==1.35
urllib3==1.25.8
uvicorn==0.11.3
uvloop==0.14.0
websockets==8.1
XlsxWriter~=1.3.7